requirements.txt
fastapi==0.95.2
fastapi-utils==0.2.1
uvicorn[standard]==0.22.0
langchain==0.0.181
pyjwt
pytz==2023.3
openai
tqdm==4.65.0
tiktoken
backoff==2.2.1
pymilvus==2.2.6
python-multipart==0.0.6
unstructured==0.7.10
google-cloud-secret-manager
google-cloud-bigquery
google-cloud-firestore
google-cloud-storage
db-dtypes
pytest==7.4.0
pytest-asyncio==0.21.1